+++ binance/spot/loan.py
@@ -1,0 +1,26 @@
+from binance.lib.utils import check_required_parameter
+
+
+def loan_history(self, asset: str, **kwargs):
+    """Get Crypto Loans Income History (USER_DATA)
+
+    GET /sapi/v1/loan/income
+
+    https://binance-docs.github.io/apidocs/spot/en/#get-crypto-loans-income-history-user_data
+
+    Args:
+      asset (str)
+    Keyword Args:
+      type (str, optional): All types will be returned by default.
+                            borrowIn, collateralSpent, repayAmount, collateralReturn (collateral return after repayment),
+                            addCollateral, removeCollateral, collateralReturnAfterLiquidation
+      startTime (int, optional)
+      endTime (int, optional)
+      limit (int, optional): default 20, max 100
+      recvWindow (int, optional): The value cannot be greater than 60000
+    """
+
+    check_required_parameter(asset, "asset")
+
+    payload = {"asset": asset, **kwargs}
+    return self.sign_request("GET", "/sapi/v1/loan/income", payload)